scripts/requirements.txt
# Folivm docs/RAIDD generator. Install: pip install -r scripts/requirements.txt
PyYAML>=6.0
